task2/requirements.txt
langchain
langchain-openai
python-dotenv
pydantic
pytest
